move reaper code into its own CLI tool
