+++ reaper-templates/warning1.txt
@@ -12,9 +12,12 @@ If you would like to extend the life of your AWSBOX please run this command:
 
 and I will let it live for a little while longer. 
 
+
 If this AWSBOX should NEVER be shutdown please run this command: 
 
   > awsbox NOKILL ==INSTANCE_ID==
+
+and I will ignore it forever.
 
 If you do not need it anymore please ignore this warning.
